Discovery connected and working, need to check queries
New API deprecated the testing method used in the original notebook. Followed API doc to auth and connect, and it is working.
Queries seem to have changed too, the default included in the original notebook is not working either, and there are other categories, need to keep working on how to translate them: check API doc and how to translate into the notebook code